x/boxoffice/types/message_buy_ticket_test.go
package types

import (
	"testing"

	"github.com/bermuell/boxoffice/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgBuyTicket_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgBuyTicket
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgBuyTicket{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgBuyTicket{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
